requirements.txt
transformers==4.51.1
torch==2.6.0
tqdm==4.67.1
biopython==1.84
colorama==0.4.4
safetensors==0.5.3
